Uploaded new Instructions, Added Attk anim events, fixed enemy anim controllers (see desc.)
Previously, the default state for enemy anim controllers was to go straight into their walking anim. With the recent addition of the Idle anim by benson into the controller, but without the logic behind it to support and idle, no anims played.

I have bypassed the idle anim state in the controller, so now it is similar to before, where default state is walking.